feat: add removing a few ideas

+++ ideas.py
@@ -93,18 +93,22 @@ def test_new_idea(name, desc):
 
 # TODO: обновить ux всех remove и добавить флаг по триграммам
 # TODO: обновить способ поиска в remove
-def remove_idea_by_date(date=False, time=False):
+# TODO: добавить версию идеи 
+def remove_idea_by_date(date=None, time=None):
 
     if date: date = _clear_num(date)
     if time: time = _clear_num(time)
 
     coincidences = []
     for i in _get_list_of_files(folder_ideas): # i202608041359_dd0cd3.md -> i 2026-08-04 13:59 _ dd0cd3 .md
-        if not time:
+        if time is None:
+            if date is None:
+                print("E: No Data Or Time Entered")
+                return
             if i[1:9] == date:
                 coincidences.append(i)
         else:
-            if not date:
+            if date is None:
                 if i[9:13] == time:
                     coincidences.append(i)
             else:
@@ -118,8 +122,7 @@ def remove_idea_by_date(date=False, time=False):
 
     elif len(coincidences) == 1:
         # TODO: добавить так чтобы была понятна какая идея удаляется
-        
-        # TODO: переместить data["warnings"]["ideas"]["delete"] в config
+        print(f"Found 1 idea:\ndate: {coincidences[0][1:5]}-{coincidences[0][5:7]}-{coincidences[0][7:9]}, uuid: {coincidences[0][14:20]}")
         if _warning(data["warnings"]["ideas"]["delete"]):
             print("Deleting idea...")
             (folder_ideas / coincidences[0]).unlink()
@@ -134,17 +137,36 @@ def remove_idea_by_date(date=False, time=False):
             print(f"  {i+1}. {date_coincidences} {time_coincidences} [{coincidences[i][14:-3]}]")
 
         print(data["settings"]["all"]["separator_symbol"] * data["settings"]["all"]["separator_length"])
-        try:
-            # TODO: переместить в config
-            answer = int(input(f"Enter number to delete (1-{len(coincidences)}, or 0 to cancel): "))
-        except ValueError:
+
+        # TODO: переместить в config
+        answer = input(f"Enter number to delete (1-{len(coincidences)}, 0 to cancel or enter ideas via ',' (1,2,3,4)): ")
+        answer = answer.split(",")
+        answer = answer[0] if len(answer) == 1 else answer
+        if not isinstance(answer, list) and not answer.isdigit():
             print("E: Incorrect number")
             print("Cancellation...")
             answer = 0
-        if answer != 0:
+        if answer != 0 and not isinstance(answer, list):
+            print(f"Selected idea:\ndate: {coincidences[answer-1][1:5]}-{coincidences[answer-1][5:7]}-{coincidences[answer-1][7:9]}, uuid: {coincidences[answer-1][14:20]}")
             if _warning(data["warnings"]["ideas"]["delete"]):
                 print("Deleting idea...")
                 (folder_ideas / coincidences[answer-1]).unlink()
+        if isinstance(answer, list):
+            for i in answer:
+                if not i.isdigit():
+                    print("E: Incorrect number")
+                    print("Cancellation...")
+                    return
+                if i == "0":
+                    return
+            print("Selected ideas:")
+            for i in answer:
+                print(f"date: {coincidences[int(i)-1][1:5]}-{coincidences[int(i)-1][5:7]}-{coincidences[int(i)-1][7:9]}, uuid: {coincidences[int(i)-1][14:20]}")
+
+            if _warning(data["warnings"]["ideas"]["delete"]): # FIXME: исправить текст, там this idea, надо these ideas
+                print("Deleting ideas...")
+                for i in answer:
+                    (folder_ideas / coincidences[int(i)-1]).unlink()
 
 def remove_idea_by_uuid(uuid_):
